==> R/6-compare-sero-pcr.Rmd <==
---
title: "Monitoring transmission intensity of trachoma with serology"
subtitle: "Compare trachoma indicators"
author: "Christine Tedijanto christine.tedijanto@ucsf.edu"
date: "Last updated: `r Sys.Date()`"
output:
    html_document:
      code_folding: hide
      toc: true
      toc_float: true
      toc_collapsed: true
      toc_depth: 3
number_sections: false
---

**Objective:** Estimate correlations between trachoma indicators and visualize.

```{r setup, include=FALSE}
#------------------------------
# Load project config file
#------------------------------
library(here)
source(here("R/0-config.R"))

#------------------------------
# global variables
#------------------------------

# minimum number of measurements per cluster
# see 1-trachoma-sero-transmission-make-datasets.Rmd
# for examination of minimum cluster size. 15 chosen
min_clu_n <- 15 

# save figures?
save_figs <- TRUE 
```

# Load data

```{r read data}

#------------------------------
# load indiv level data to
# estimate the SCR among 1-5y
# for comparisons below (not
# completed in main SCR script)
#------------------------------
ind_df <- read_rds(here("data","trachoma_sero_transmission_analysis_indiv.rds")) %>%
  filter(pgp3_minobs) %>%
  filter(!is.na(pgp3_pos))

#------------------------------
# load cluster-level estimates
# these estimates were generated
# in the script 2-estimate-foi.Rmd
#------------------------------
clu_all  <- read_rds(here("output", "clu_all_summary.rds")) 
clu_1to5 <- read_rds(here("output", "clu_1to5_summary.rds")) 
lyn_all  <- read_rds(here("output", "lyn_all_summary.rds"))

#------------------------------
# encode Carter Center overall
# study-level PCR prevalence
# estimates
# and join to the location-year-name (lyn)
# data
#------------------------------
cc_pcr_manual <- data.frame(location_year_name = c("Andabet, Ethiopia 2017",
                                                   "Dera, Ethiopia 2017",
                                                   "Woreta town, Ethiopia 2017",
                                                   "Alefa, Ethiopia 2017"),
                            prev = c(11.3, 0, 0, 0),
                            stringsAsFactors = FALSE)

lyn_all <- lyn_all %>%
  left_join(cc_pcr_manual, by = "location_year_name") %>% 
  mutate(pcr_prev_median = coalesce(prev, pcr_prev_median)) %>% 
  dplyr::select(-prev)

#------------------------------
# create a cluster-level dataset
# restricted to pops with 1-9y
# for convenience
# two studies did not measure
# PCR in children 6-9 y
# so they are excluded here
#------------------------------
clu_1to9 <- clu_all %>%
  filter(!location_year_name %in% c("Wag Hemra, Ethiopia (TAITU) 2018",
                                   "Matameye, Niger (PRET) 2013")
         )

```

# Estimate study-level SCR among 1-5y

These estimates were not created at the study level in the main force of infection (FOI) estimation script, `2-estimate-foi.Rmd`, so add them here.

```{r estimate scr among 1 to 5}
#------------------------------
# loop over studies and 
# estimate SCR using an SIR
# model with a constant rate
#------------------------------
scr_ests_1to5 <- foreach(lyni = unique(ind_df$location_year_name), .combine = rbind) %do% {
  di <- ind_df %>%
    filter(location_year_name == lyni & age_years >=1 & age_years <= 5)
  glm_fit <- glm(pgp3_pos~1, offset = log(age_years), data = di, family = binomial(link = "cloglog"))
  lambda <- as.numeric(exp(glm_fit$coefficients))
  return(data.frame(location_year_name = lyni, sir_foi_1to5 = lambda)) 
}

#------------------------------
# merge estimates to the 
# study level data
# subset to o
#------------------------------
lyn_all <- lyn_all %>%
  left_join(scr_ests_1to5, by = "location_year_name")
```

```{r stratify estimates for taitu}
#------------------------------
# for TAITU2018, there are 
# clusters that are MDA and 
# no MDA in the RCT. Create
# study level summaries of these
# two subgroups and replace the 
# overall means in the study-level file
# for stratified estimates and
# plotting of study-level values
# in the MDA-stratified analyses
#------------------------------
taitu2018 <- ind_df %>% 
  filter(study_id == "TAITU2018")
# PCR, seroprev, and TF prev
taitu2018_bymda <- taitu2018 %>%
  group_by(study_id,location_year_name,mda,cluster_id) %>%
  summarize(pgp3_prev = mean(pgp3_pos, na.rm = TRUE),
            pcr_prev = mean(pcr, na.rm = TRUE), 
            tf_prev = mean(tf, na.rm = TRUE),
            .groups = "drop") %>%
  group_by(study_id,location_year_name,mda) %>%
  summarize(pgp3_prev_median = median(pgp3_prev)*100,
            pcr_prev_median = median(pcr_prev)*100,
            tf_prev_median = median(tf_prev)*100,
            .groups = "drop")
# SCRs
taitu2018_fit_mda <- glm(pgp3_pos~1, offset = log(age_years), data = taitu2018 %>% filter(mda==1), family = binomial(link = "cloglog"))
taitu2018_fit_nomda <- glm(pgp3_pos~1, offset = log(age_years), data = taitu2018 %>% filter(mda==0), family = binomial(link = "cloglog"))
taitu2018_scr_mda <- as.numeric(exp(taitu2018_fit_mda$coefficients))
taitu2018_scr_nomda <- as.numeric(exp(taitu2018_fit_nomda$coefficients))
taitu2018_bymda <- taitu2018_bymda %>%
  mutate(sir_foi_1to5_recode = ifelse(mda == 0, taitu2018_scr_nomda, taitu2018_scr_mda)) %>%
  rename(mda_recode = mda,
         pgp3_prev_recode = pgp3_prev_median,
         pcr_prev_recode = pcr_prev_median,
         tf_prev_recode = tf_prev_median)

lyn_mda <- lyn_all %>%
  left_join(taitu2018_bymda, by = c("study_id","location_year_name")) %>%
  mutate(
    mda = ifelse(study_id == "TAITU2018", mda_recode, mda),
    pgp3_prev_median = ifelse(study_id == "TAITU2018", pgp3_prev_recode, pgp3_prev_median),
    pcr_prev_median = ifelse(study_id == "TAITU2018", pcr_prev_recode, pcr_prev_median),
    tf_prev_median = ifelse(study_id == "TAITU2018", tf_prev_recode, tf_prev_median),
    sir_foi_1to5 = ifelse(study_id == "TAITU2018", sir_foi_1to5_recode, sir_foi_1to5)
         )

  

```


# Estimate correlations

Use pearson correlation ($r$)

For inference of cluster level correlation, bootstrap resample clusters with replacement, stratified by study.

For interference of study level correlation, use the Pearson product moment estimator, implemented in `cor.test()`.

```{r correlation bootstrap function}
#------------------------------
# function to stratify by study
# and bootstrap resample clusters
# used w/in foreach below
# assumes a data frame with a
# variable location_year_name
# used to stratify, and then 
# estimate correlation
# between columns 2 and 3
#------------------------------
boot_cor <- function(df,breps=1000) {
  boot_ests <- foreach(booti = 1:breps, .combine = rbind) %dopar% {
    set.seed(booti)
    di <- df %>%
    group_by(location_year_name) %>%
    slice_sample(prop = 1, replace = TRUE)
    cori <- cor(di[,2],di[,3],method = "pearson")
    return(cori)
  }
  return(boot_ests)
}

#------------------------------
# formatting functions for
# correlation ests and 95% CIs
#------------------------------
cl_corlab <- function(corest,corlb,corub) {
  paste0("Cluster-level r (95% CI): ", sprintf("%1.2f",corest), " (", sprintf("%1.2f",corlb), ",",sprintf("%1.2f",corub), ")")
}

lyn_corlab <- function(corest,corlb,corub) {
  paste0("Population-level r (95% CI): ", sprintf("%1.2f",corest), " (", sprintf("%1.2f",corlb), ",",sprintf("%1.2f",corub), ")")
}

#------------------------------
# plotting function for
# cluster level values and
# study level values
#------------------------------
corplot <- function(dcor,dcor2,
                     xvar,xvar2,xlab,xlim=100,
                     yvar,yvar2,ylab,
                     corlab,corlab2) {
  # dcor  : dataset for cluster level values
  # dcor2 : dataset for study level values
  # xvar  : x variable in the cluster level dataset
  # xvar2 : x variable in the study level dataset
  # xlab  : label for x-axis
  # xlim  : upper limit for x-axis
  # yvar  : y variable in the cluster level dataset
  # yvar2 : y variable in the study level dataset
  # ylab  : label for y-axis
  # corlab: formatted correlation and 95% CI to print for cluster level
  # corlab2 : formatted correlation and 95% CI to print for study level
  
  ggplot() +
  geom_abline(intercept=0,slope=1,linetype = "dotted") + 
  geom_point(data = dcor , aes(x = get(xvar), y = get(yvar), color = location_year_name),
               alpha = 0.5, size = 1,
               show.legend = FALSE) +
    scale_fill_manual(values = location_year_name_colors) +
    scale_color_manual(values = location_year_name_colors) +
    labs(x = xlab, y = ylab) +
    theme_minimal() +
    lims(x = c(0, xlim), y = c(0,100)) +
    theme(axis.title = element_text(size = 10),
          axis.text = element_text(size = 9),
          legend.position = "none",
          legend.title = element_blank(),
          plot.margin = margin(t=16.5,r=5.5,b=5.5,l=5.5,unit="pt")
          ) + 
  geom_label(data = data.frame(x = 1, y = 1),
              label = corlab,
              x = 0, y = 97, size = 2, color = "black", hjust = 0, fill = "white", alpha = 0.8,
              label.size = NA) +
  geom_point(data = dcor2,
               aes(x = get(xvar2), y = get(yvar2),
                   fill = location_year_name),
               color = "black", pch = 21, alpha = 0.9,
               size = 2.5) +
      geom_label(data = data.frame(x = 1, y = 1),
              label = corlab2,
              x = 0, y = 88, size = 2, color = "black", hjust = 0, fill = "white", alpha = 0.8,
              label.size = NA)
  
}

#------------------------------
# plotting function for
# cluster level values and
# study level values
# slight variation for SCR:
# take away 1:1 line
#------------------------------
corplot_scr <- function(dcor,dcor2,
                     xvar,xvar2,xlab,xlim=100,
                     yvar,yvar2,ylab,
                     corlab,corlab2) {
  # dcor  : dataset for cluster level values
  # dcor2 : dataset for study level values
  # xvar  : x variable in the cluster level dataset
  # xvar2 : x variable in the study level dataset
  # xlab  : label for x-axis
  # xlim  : upper limit for x-axis
  # yvar  : y variable in the cluster level dataset
  # yvar2 : y variable in the study level dataset
  # ylab  : label for y-axis
  # corlab: formatted correlation and 95% CI to print for cluster level
  # corlab2 : formatted correlation and 95% CI to print for study level
  
  ggplot() +
  # geom_abline(intercept=0,slope=1,linetype = "dotted") + 
  geom_point(data = dcor , aes(x = get(xvar), y = get(yvar), color = location_year_name),
               alpha = 0.5, size = 1,
               show.legend = FALSE) +
    scale_fill_manual(values = location_year_name_colors) +
    scale_color_manual(values = location_year_name_colors) +
    labs(x = xlab, y = ylab) +
    theme_minimal() +
    lims(x = c(0, xlim), y = c(0,100)) +
    theme(axis.title = element_text(size = 10),
          axis.text = element_text(size = 9),
          legend.position = "none",
          legend.title = element_blank(),
          plot.margin = margin(t=16.5,r=5.5,b=5.5,l=5.5,unit="pt")
          ) + 
  geom_label(data = data.frame(x = 1, y = 1),
              label = corlab,
              x = 0, y = 97, size = 2, color = "black", hjust = 0, fill = "white", alpha = 0.8,
              label.size = NA) +
  geom_point(data = dcor2,
               aes(x = get(xvar2), y = get(yvar2),
                   fill = location_year_name),
               color = "black", pch = 21, alpha = 0.9,
               size = 2.5) +
      geom_label(data = data.frame(x = 1, y = 1),
              label = corlab2,
              x = 0, y = 88, size = 2, color = "black", hjust = 0, fill = "white", alpha = 0.8,
              label.size = NA)
  
}
```

## PCR vs seroprevalence
```{r estimate cluster correlations pcr vs seroprev}
#------------------------------
# PCR vs seroprev
# all ages
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, pgp3_prev)
cor_pcr_sp_clu  <- cor(dcor$pcr_prev, dcor$pgp3_prev, method = "pearson")
boot_pcr_sp_clu <- boot_cor(dcor)
cor_pcr_sp_clu_lb <- quantile(boot_pcr_sp_clu[,1],probs=0.025)
cor_pcr_sp_clu_ub <- quantile(boot_pcr_sp_clu[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_sp_clu,cor_pcr_sp_clu_lb,cor_pcr_sp_clu_ub)
# study level correlation
dcor2 <- lyn_all %>% 
  select(location_year_name,pcr_prev_median, pgp3_prev_median) %>%
  filter(!is.na(pcr_prev_median))
cor_pcr_sp_lyn <- cor.test(dcor2$pcr_prev_median, dcor2$pgp3_prev_median)
corlab2 <- lyn_corlab(cor_pcr_sp_lyn$estimate, cor_pcr_sp_lyn$conf.int[1],cor_pcr_sp_lyn$conf.int[2])
# make summary figure
sp_all_fig <- corplot(
  dcor,dcor2,
  xvar = "pgp3_prev", xvar2 = "pgp3_prev_median", xlab = "Pgp3 seroprevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("All") + theme(plot.title = element_text(face = "plain", size=12))


#------------------------------
# PCR vs seroprev
# 1-9y
#------------------------------
# cluster level correlation
dcor <- clu_1to9 %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, pgp3_prev)
cor_pcr_sp_clu_1to9  <- cor(dcor$pcr_prev, dcor$pgp3_prev, method = "pearson")
boot_pcr_sp_clu_1to9 <- boot_cor(dcor)
cor_pcr_sp_clu_1to9_lb <- quantile(boot_pcr_sp_clu_1to9[,1],probs=0.025)
cor_pcr_sp_clu_1to9_ub <- quantile(boot_pcr_sp_clu_1to9[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_sp_clu_1to9,cor_pcr_sp_clu_1to9_lb,cor_pcr_sp_clu_1to9_ub)

# study level correlation
dcor2 <- lyn_all %>%
  # drop 3 studies with no PCR measurements 6-9y
  filter(!study_id %in% c("TAITU2018","PRET2013","TCC-Ethiopia2017")) %>%
  select(location_year_name, pcr_prev_median, pgp3_prev_median) %>%
  filter(!is.na(pcr_prev_median) & !is.na(pgp3_prev_median))
cor_pcr_sp_lyn_1to9 <- cor.test(dcor2$pcr_prev_median, dcor2$pgp3_prev_median)
corlab2 <- lyn_corlab(cor_pcr_sp_lyn_1to9$estimate, cor_pcr_sp_lyn_1to9$conf.int[1],cor_pcr_sp_lyn_1to9$conf.int[2])

# make summary figure
sp_1to9_fig <- corplot(
  dcor,dcor2,
  xvar = "pgp3_prev", xvar2 = "pgp3_prev_median", xlab = "Pgp3 seroprevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("1 to 9 year olds") + theme(plot.title = element_text(face = "plain", size=12))


#------------------------------
# PCR vs seroprev
# 1-5y 
#------------------------------
# cluster level correlation
dcor <- clu_1to5 %>% 
  filter(pcr_n_tested_1to5y >= min_clu_n & pgp3_n_tested_1to5y >= min_clu_n) %>%
  select(location_year_name, pcr_prev_1to5y, pgp3_prev_1to5y)
cor_pcr_sp_clu_1to5  <- cor(dcor$pcr_prev_1to5y, dcor$pgp3_prev_1to5y, method = "pearson")
boot_pcr_sp_clu_1to5 <- boot_cor(dcor)
cor_pcr_sp_clu_1to5_lb <- quantile(boot_pcr_sp_clu_1to5[,1],probs=0.025)
cor_pcr_sp_clu_1to5_ub <- quantile(boot_pcr_sp_clu_1to5[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_sp_clu_1to5,cor_pcr_sp_clu_1to5_lb,cor_pcr_sp_clu_1to5_ub)
# study level correlation
# since there is not study-level
# summary of 1-5y, need to compute
# medians first
# also need to add-in the TCC
# studies, which have study-level
# PCR only (but no cluster-level)
dcor2 <- clu_1to5 %>%
  select(study_id,location_year_name,pcr_prev, pgp3_prev) %>%
  group_by(study_id,location_year_name) %>%
  summarize(pcr_prev_median = median(pcr_prev),
            pgp3_prev_median = median(pgp3_prev), 
            .groups = "drop"
            ) %>%
  # add in the TCC ests
  left_join(lyn_all %>% select(location_year_name, pcr_prev_med = pcr_prev_median), by = "location_year_name") %>%
  mutate(pcr_prev_median = ifelse(study_id=="TCC-Ethiopia2017",pcr_prev_med, pcr_prev_median)) %>%
  filter(!is.na(pcr_prev_median),!is.na(pgp3_prev_median))

cor_pcr_sp_lyn_1to5 <- cor.test(dcor2$pcr_prev_median, dcor2$pgp3_prev_median)
corlab2 <- lyn_corlab(cor_pcr_sp_lyn_1to5$estimate, cor_pcr_sp_lyn_1to5$conf.int[1],cor_pcr_sp_lyn_1to5$conf.int[2])

# make summary figure
sp_1to5_fig <- corplot(
  dcor, dcor2,
  xvar = "pgp3_prev_1to5y", xvar2 = "pgp3_prev_median", xlab = "Pgp3 seroprevalence (%)",xlim=100,
  yvar = "pcr_prev_1to5y",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("1 to 5 year olds") + theme(plot.title = element_text(face = "plain", size=12))


#------------------------------
# PCR vs seroprev
# MDA
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n & mda==1) %>%
  select(location_year_name, pcr_prev, pgp3_prev)
cor_pcr_sp_clu_mda  <- cor(dcor$pcr_prev, dcor$pgp3_prev, method = "pearson")
boot_pcr_sp_clu_mda <- boot_cor(dcor)
cor_pcr_sp_clu_mda_lb <- quantile(boot_pcr_sp_clu_mda[,1],probs=0.025)
cor_pcr_sp_clu_mda_ub <- quantile(boot_pcr_sp_clu_mda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_sp_clu_mda,cor_pcr_sp_clu_mda_lb,cor_pcr_sp_clu_mda_ub)
# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(pgp3_prev_median) & mda == 1) %>%
  select(location_year_name, pcr_prev_median, pgp3_prev_median)
cor_pcr_sp_lyn_mda <- cor.test(dcor2$pcr_prev_median, dcor2$pgp3_prev_median)
corlab2 <- lyn_corlab(cor_pcr_sp_lyn_mda$estimate, cor_pcr_sp_lyn_mda$conf.int[1],cor_pcr_sp_lyn_mda$conf.int[2])

# make summary figure
sp_MDA_fig <- corplot(
  dcor,dcor2,
  xvar = "pgp3_prev", xvar2 = "pgp3_prev_median", xlab = "Pgp3 seroprevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("MDA in past year") + theme(plot.title = element_text(face = "plain", size=12))



#------------------------------
# PCR vs seroprev
# no MDA
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n & mda==0) %>%
  select(location_year_name, pcr_prev, pgp3_prev)
cor_pcr_sp_clu_nomda  <- cor(dcor$pcr_prev, dcor$pgp3_prev, method = "pearson")
boot_pcr_sp_clu_nomda <- boot_cor(dcor)
cor_pcr_sp_clu_nomda_lb <- quantile(boot_pcr_sp_clu_nomda[,1],probs=0.025)
cor_pcr_sp_clu_nomda_ub <- quantile(boot_pcr_sp_clu_nomda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_sp_clu_nomda,cor_pcr_sp_clu_nomda_lb,cor_pcr_sp_clu_nomda_ub)

# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(pgp3_prev_median) & mda == 0) %>%
  select(location_year_name, pcr_prev_median, pgp3_prev_median)
cor_pcr_sp_lyn_nomda <- cor.test(dcor2$pcr_prev_median, dcor2$pgp3_prev_median)
corlab2 <- lyn_corlab(cor_pcr_sp_lyn_nomda$estimate, cor_pcr_sp_lyn_nomda$conf.int[1],cor_pcr_sp_lyn_mda$conf.int[2])


# make summary figure
sp_noMDA_fig <- corplot(
  dcor,dcor2,
  xvar = "pgp3_prev", xvar2 = "pgp3_prev_median", xlab = "Pgp3 seroprevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("No MDA in past year") + theme(plot.title = element_text(face = "plain", size=12))

```



## PCR vs SCR
```{r estimate cluster correlations pcr vs scr}
#------------------------------
# PCR vs SCR
# all ages
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_clu  <- cor(dcor$pcr_prev, dcor$sir_foi, method = "pearson")
boot_pcr_scr_clu <- boot_cor(dcor)
cor_pcr_scr_clu_lb <- quantile(boot_pcr_scr_clu[,1],probs=0.025)
cor_pcr_scr_clu_ub <- quantile(boot_pcr_scr_clu[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_scr_clu,cor_pcr_scr_clu_lb,cor_pcr_scr_clu_ub)

# study level correlation
dcor2 <- lyn_all %>% 
  select(location_year_name, pcr_prev_median, sir_foi) %>%
  filter(!is.na(pcr_prev_median) & !is.na(sir_foi)) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_lyn <- cor.test(dcor2$pcr_prev_median, dcor2$sir_foi)
corlab2 <- lyn_corlab(cor_pcr_scr_lyn$estimate, cor_pcr_scr_lyn$conf.int[1],cor_pcr_scr_lyn$conf.int[2])

# make summary figure
scr_all_fig <- corplot_scr(
  dcor,dcor2,
  xvar = "sir_foi", xvar2 = "sir_foi", xlab = "Seroconversion rate per 100 PY",xlim=50,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("All") + theme(plot.title = element_text(face = "plain", size=12))

#------------------------------
# PCR vs SCR
# 1-9y
#------------------------------
# cluster level correlation
dcor <- clu_1to9 %>% 
  # drop 3 studies with no PCR measurements 6-9y
  filter(!study_id %in% c("TAITU2018","PRET2013","TCC-Ethiopia2017")) %>%
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)

cor_pcr_scr_clu_1to9  <- cor(dcor$pcr_prev, dcor$sir_foi, method = "pearson")
boot_pcr_scr_clu_1to9 <- boot_cor(dcor)
cor_pcr_scr_clu_1to9_lb <- quantile(boot_pcr_scr_clu_1to9[,1],probs=0.025)
cor_pcr_scr_clu_1to9_ub <- quantile(boot_pcr_scr_clu_1to9[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_scr_clu_1to9,cor_pcr_scr_clu_1to9_lb,cor_pcr_scr_clu_1to9_ub)

# study level correlation
dcor2 <- lyn_all %>%
  # drop 3 studies with no PCR measurements 6-9y
  filter(!study_id %in% c("TAITU2018","PRET2013","TCC-Ethiopia2017")) %>%
  select(location_year_name, pcr_prev_median, sir_foi) %>%
  filter(!is.na(pcr_prev_median) & !is.na(sir_foi)) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_lyn_1to9 <- cor.test(dcor2$pcr_prev_median, dcor2$sir_foi)
corlab2 <- lyn_corlab(cor_pcr_scr_lyn_1to9$estimate, cor_pcr_scr_lyn_1to9$conf.int[1],cor_pcr_scr_lyn_1to9$conf.int[2])

# make summary figure
scr_1to9_fig <- corplot_scr(
  dcor,dcor2,
  xvar = "sir_foi", xvar2 = "sir_foi", xlab = "Seroconversion rate per 100 PY",xlim=50,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("1 to 9 year olds") + theme(plot.title = element_text(face = "plain", size=12))


#------------------------------
# PCR vs SCR
# 1-5y 
#------------------------------
# cluster level correlation
dcor <- clu_1to5 %>% 
  filter(pcr_n_tested_1to5y >= min_clu_n & pgp3_n_tested_1to5y >= min_clu_n) %>%
  select(location_year_name, pcr_prev_1to5y, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_clu_1to5  <- cor(dcor$pcr_prev_1to5y, dcor$sir_foi, method = "pearson")
boot_pcr_scr_clu_1to5 <- boot_cor(dcor)
cor_pcr_scr_clu_1to5_lb <- quantile(boot_pcr_scr_clu_1to5[,1],probs=0.025)
cor_pcr_scr_clu_1to5_ub <- quantile(boot_pcr_scr_clu_1to5[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_scr_clu_1to9,cor_pcr_scr_clu_1to9_lb,cor_pcr_scr_clu_1to9_ub)

# study level correlation
# since there is not study-level
# summary of 1-5y, need to compute
# medians first
# also need to add-in the TCC
# studies, which have study-level
# PCR only (but no cluster-level)
dcor2 <- clu_1to5 %>%
  select(study_id,location_year_name,pcr_prev) %>%
  group_by(study_id,location_year_name) %>%
  summarize(pcr_prev_median = median(pcr_prev), .groups = "drop") %>%
  # add in the TCC ests for PCR and the study-level SCR for 1-5y
  left_join(lyn_all %>% select(location_year_name, pcr_prev_med = pcr_prev_median, sir_foi_1to5),by = "location_year_name") %>%
  mutate(pcr_prev_median = ifelse(study_id=="TCC-Ethiopia2017",pcr_prev_med, pcr_prev_median),
         sir_foi_1to5 = sir_foi_1to5*100) %>%
  filter(!is.na(pcr_prev_median))
cor_pcr_scr_lyn_1to5 <- cor.test(dcor2$pcr_prev_median, dcor2$sir_foi_1to5)
corlab2 <- lyn_corlab(cor_pcr_scr_lyn_1to5$estimate, cor_pcr_scr_lyn_1to5$conf.int[1],cor_pcr_scr_lyn_1to5$conf.int[2])

# make summary figure
scr_1to5_fig <- corplot_scr(
  dcor,dcor2,
  xvar = "sir_foi", xvar2 = "sir_foi_1to5", xlab = "Seroconversion rate per 100 PY",xlim=50,
  yvar = "pcr_prev_1to5y",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("1 to 5 year olds") + theme(plot.title = element_text(face = "plain", size=12))



#------------------------------
# PCR vs SCR
# MDA
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n & mda==1) %>%
  select(location_year_name, pcr_prev, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_clu_mda  <- cor(dcor$pcr_prev, dcor$sir_foi, method = "pearson")
boot_pcr_scr_clu_mda <- boot_cor(dcor)
cor_pcr_scr_clu_mda_lb <- quantile(boot_pcr_scr_clu_mda[,1],probs=0.025)
cor_pcr_scr_clu_mda_ub <- quantile(boot_pcr_scr_clu_mda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_scr_clu_mda,cor_pcr_scr_clu_mda_lb,cor_pcr_scr_clu_mda_ub)

# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(sir_foi) & mda == 1) %>%
  select(location_year_name, pcr_prev_median, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_lyn_mda <- cor.test(dcor2$pcr_prev_median, dcor2$sir_foi)
corlab2 <- lyn_corlab(cor_pcr_scr_lyn_mda$estimate, cor_pcr_scr_lyn_mda$conf.int[1],cor_pcr_scr_lyn_mda$conf.int[2])

# make summary figure
scr_MDA_fig <- corplot_scr(
  dcor,dcor2,
  xvar = "sir_foi", xvar2 = "sir_foi", xlab = "Seroconversion rate per 100 PY",xlim=50,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("MDA in past year") + theme(plot.title = element_text(face = "plain", size=12))



#------------------------------
# PCR vs SCR
# no MDA
#------------------------------
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & pgp3_n_tested >= min_clu_n & mda==0) %>%
  select(location_year_name, pcr_prev, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_clu_nomda  <- cor(dcor$pcr_prev, dcor$sir_foi, method = "pearson")
boot_pcr_scr_clu_nomda <- boot_cor(dcor)
cor_pcr_scr_clu_nomda_lb <- quantile(boot_pcr_scr_clu_nomda[,1],probs=0.025)
cor_pcr_scr_clu_nomda_ub <- quantile(boot_pcr_scr_clu_nomda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_scr_clu_nomda,cor_pcr_scr_clu_nomda_lb,cor_pcr_scr_clu_nomda_ub)

# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(sir_foi) & mda == 0) %>%
  select(location_year_name, pcr_prev_median, sir_foi) %>%
  mutate(sir_foi = sir_foi*100)
cor_pcr_scr_lyn_nomda <- cor.test(dcor2$pcr_prev_median, dcor2$sir_foi)
corlab2 <- lyn_corlab(cor_pcr_scr_lyn_nomda$estimate, cor_pcr_scr_lyn_nomda$conf.int[1],cor_pcr_scr_lyn_nomda$conf.int[2])

# make summary figure
scr_noMDA_fig <- corplot_scr(
  dcor,dcor2,
  xvar = "sir_foi", xvar2 = "sir_foi", xlab = "Seroconversion rate per 100 PY",xlim=50,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  ) +
  ggtitle("No MDA in past year") + theme(plot.title = element_text(face = "plain", size=12))


```

## PCR vs TF prevalence
```{r estimate cluster correlations pcr vs tf}
#------------------------------
# PCR vs TF
# all ages
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & tf_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, tf_prev)
cor_pcr_tf_clu  <- cor(dcor$pcr_prev, dcor$tf_prev, method = "pearson")
boot_pcr_tf_clu <- boot_cor(dcor)
cor_pcr_tf_clu_lb <- quantile(boot_pcr_tf_clu[,1],probs=0.025)
cor_pcr_tf_clu_ub <- quantile(boot_pcr_tf_clu[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_tf_clu,cor_pcr_tf_clu_lb,cor_pcr_tf_clu_ub)

# study level correlation
dcor2 <- lyn_all %>% 
      select(location_year_name,pcr_prev_median, tf_prev_median) %>%
      filter(!is.na(pcr_prev_median) & !is.na(tf_prev_median))
cor_pcr_tf_lyn <- cor.test(dcor2$pcr_prev_median, dcor2$tf_prev_median)
corlab2 <- lyn_corlab(cor_pcr_tf_lyn$estimate, cor_pcr_tf_lyn$conf.int[1],cor_pcr_tf_lyn$conf.int[2])

# make summary figure
tf_all_fig <- corplot(
  dcor,dcor2,
  xvar = "tf_prev", xvar2 = "tf_prev_median", xlab = "TF prevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  )


#------------------------------
# PCR vs TF
# 1-9y
#------------------------------
# cluster level correlation
dcor <- clu_1to9 %>% 
  # drop 3 studies with no PCR measurements 6-9y
  filter(!study_id %in% c("TAITU2018","PRET2013","TCC-Ethiopia2017")) %>%
  filter(pcr_n_tested >= min_clu_n & tf_n_tested >= min_clu_n) %>%
  select(location_year_name, pcr_prev, tf_prev)
cor_pcr_tf_clu_1to9  <- cor(dcor$pcr_prev, dcor$tf_prev, method = "pearson")
boot_pcr_tf_clu_1to9 <- boot_cor(dcor)
cor_pcr_tf_clu_1to9_lb <- quantile(boot_pcr_tf_clu_1to9[,1],probs=0.025)
cor_pcr_tf_clu_1to9_ub <- quantile(boot_pcr_tf_clu_1to9[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_tf_clu_1to9,cor_pcr_tf_clu_1to9_lb,cor_pcr_tf_clu_1to9_ub)

# study level correlation
dcor2 <- lyn_all %>%
  # drop 3 studies with no PCR measurements 6-9y
  filter(!study_id %in% c("TAITU2018","PRET2013","TCC-Ethiopia2017")) %>%
  select(location_year_name, pcr_prev_median, tf_prev_median) %>%
  filter(!is.na(pcr_prev_median) & !is.na(tf_prev_median))
cor_pcr_tf_lyn_1to9 <- cor.test(dcor2$pcr_prev_median, dcor2$tf_prev_median)
corlab2 <- lyn_corlab(cor_pcr_tf_lyn_1to9$estimate, cor_pcr_tf_lyn_1to9$conf.int[1],cor_pcr_tf_lyn_1to9$conf.int[2])

# make summary figure
tf_1to9_fig <- corplot(
  dcor,dcor2,
  xvar = "tf_prev", xvar2 = "tf_prev_median", xlab = "TF prevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  )


#------------------------------
# PCR vs TF
# 1-5y 
#------------------------------
# cluster level correlation
dcor <- clu_1to5 %>% 
  filter(pcr_n_tested_1to5y >= min_clu_n & tf_n_tested_1to5y >= min_clu_n) %>%
  select(location_year_name, pcr_prev_1to5y, tf_prev_1to5y)
cor_pcr_tf_clu_1to5  <- cor(dcor$pcr_prev_1to5y, dcor$tf_prev_1to5y, method = "pearson")
boot_pcr_tf_clu_1to5 <- boot_cor(dcor)
cor_pcr_tf_clu_1to5_lb <- quantile(boot_pcr_tf_clu_1to5[,1],probs=0.025)
cor_pcr_tf_clu_1to5_ub <- quantile(boot_pcr_tf_clu_1to5[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_tf_clu_1to5,cor_pcr_tf_clu_1to5_lb,cor_pcr_tf_clu_1to5_ub)

# study level correlation
# since there is not study-level
# summary of 1-5y, need to compute
# medians first
dcor2 <- clu_1to5 %>%
  select(study_id,location_year_name,pcr_prev_1to5y, tf_prev_1to5y) %>%
  group_by(study_id, location_year_name) %>%
  summarize(pcr_prev_median = median(pcr_prev_1to5y),
            tf_prev_median = median(tf_prev_1to5y),
            .groups = "drop") %>%
  # add in the study-level TCC ests (although we actually do not have TF prev for 1-5y in these studies so dropped)
  left_join(lyn_all %>% select(location_year_name, pcr_prev_med = pcr_prev_median), by = "location_year_name") %>%
  mutate(pcr_prev_median = ifelse(study_id=="TCC-Ethiopia2017",pcr_prev_med, pcr_prev_median)) %>%
  filter(!is.na(pcr_prev_median) & !is.na(tf_prev_median))

cor_pcr_tf_lyn_1to5 <- cor.test(dcor2$pcr_prev_median, dcor2$tf_prev_median)
corlab2 <- lyn_corlab(cor_pcr_tf_lyn_1to5$estimate, cor_pcr_tf_lyn_1to5$conf.int[1],cor_pcr_tf_lyn_1to5$conf.int[2])

# make summary figure
tf_1to5_fig <- corplot(
  dcor,dcor2,
  xvar = "tf_prev_1to5y", xvar2 = "tf_prev_median", xlab = "TF prevalence (%)",xlim=100,
  yvar = "pcr_prev_1to5y",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  )

#------------------------------
# PCR vs TF
# MDA
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & tf_n_tested >= min_clu_n & mda==1) %>%
  select(location_year_name, pcr_prev, tf_prev)
cor_pcr_tf_clu_mda  <- cor(dcor$pcr_prev, dcor$tf_prev, method = "pearson")
boot_pcr_tf_clu_mda <- boot_cor(dcor)
cor_pcr_tf_clu_mda_lb <- quantile(boot_pcr_tf_clu_mda[,1],probs=0.025)
cor_pcr_tf_clu_mda_ub <- quantile(boot_pcr_tf_clu_mda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_tf_clu_mda,cor_pcr_tf_clu_mda_lb,cor_pcr_tf_clu_mda_ub)

# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(tf_prev_median) & mda == 1) %>%
  select(location_year_name, pcr_prev_median, tf_prev_median)
cor_pcr_tf_lyn_mda <- cor.test(dcor2$pcr_prev_median, dcor2$tf_prev_median)
corlab2 <- lyn_corlab(cor_pcr_tf_lyn_mda$estimate, cor_pcr_tf_lyn_mda$conf.int[1],cor_pcr_tf_lyn_mda$conf.int[2])

# make summary figure
tf_MDA_fig <- corplot(
  dcor,dcor2,
  xvar = "tf_prev", xvar2 = "tf_prev_median", xlab = "TF prevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  )


#------------------------------
# PCR vs TF
# no MDA
#------------------------------
# cluster level correlation
dcor <- clu_all %>% 
  filter(pcr_n_tested >= min_clu_n & tf_n_tested >= min_clu_n & mda==0) %>%
  select(location_year_name, pcr_prev, tf_prev)
cor_pcr_tf_clu_nomda  <- cor(dcor$pcr_prev, dcor$tf_prev, method = "pearson")
boot_pcr_tf_clu_nomda <- boot_cor(dcor)
cor_pcr_tf_clu_nomda_lb <- quantile(boot_pcr_tf_clu_nomda[,1],probs=0.025)
cor_pcr_tf_clu_nomda_ub <- quantile(boot_pcr_tf_clu_nomda[,1],probs=0.975)
corlab <- cl_corlab(cor_pcr_tf_clu_nomda,cor_pcr_tf_clu_nomda_lb,cor_pcr_tf_clu_nomda_ub)

# study level correlation
dcor2 <- lyn_mda %>% 
  filter(!is.na(pcr_prev_median) & !is.na(tf_prev_median) & mda == 0) %>%
  select(location_year_name, pcr_prev_median, tf_prev_median)
cor_pcr_tf_lyn_nomda <- cor.test(dcor2$pcr_prev_median, dcor2$tf_prev_median)
corlab2 <- lyn_corlab(cor_pcr_tf_lyn_nomda$estimate, cor_pcr_tf_lyn_nomda$conf.int[1],cor_pcr_tf_lyn_nomda$conf.int[2])

# make summary figure
tf_noMDA_fig <- corplot(
  dcor,dcor2,
  xvar = "tf_prev", xvar2 = "tf_prev_median", xlab = "TF prevalence (%)",xlim=100,
  yvar = "pcr_prev",  yvar2 = "pcr_prev_median",  ylab = "PCR prevalence (%)",
  corlab,corlab2
  )


```


### Create legend

```{r legend}
# extract legend
fig4_lyn <- lyn_all %>%
  filter(!is.na(pcr_prev_median)) %>%
  pull(location_year_name)


title_right <- ""

my_ggp_right <- ggplot(data = lyn_all %>%
                     filter(location_year_name %in% fig4_lyn)) +
  geom_point(aes(x = pgp3_prev_median, y = sir_foi_median, fill = location_year_name),
             pch = 21, color = "black", size = 2.5) +
  scale_fill_manual(values = location_year_name_colors[fig4_lyn],
                    labels = location_year_name_short %>% filter(location_year_name %in% fig4_lyn) %>% deframe()) +
  theme_minimal() +
  labs(fill = title_right) +
  theme(legend.title = element_text(face = "bold", size = 9),
        legend.key.height = unit(0.5, "cm"),
        legend.key.width = unit(0.6, "cm"),
        legend.text = element_text(size = 7),
        legend.title.align=0.5) +
  guides(fill = guide_legend(nrow = 3))

step1 <- ggplot_gtable(ggplot_build(my_ggp_right))
step2 <- which(sapply(step1$grobs, function(x) x$name) == "guide-box")
fig4_step3 <- step3 <- step1$grobs[[step2]]

```

### Figure 4

```{r fig4, fig.width = 8.5, fig.height = 5.4}
fig4 <- plot_grid(plot_grid(sp_all_fig, sp_noMDA_fig, sp_MDA_fig, nrow = 1),
                  plot_grid(tf_all_fig, tf_noMDA_fig, tf_MDA_fig, nrow = 1),
                  fig4_step3,
                  ncol = 1, rel_heights = c(1, 0.9, 0.3), labels = c('a', 'b', ""))

fig4

if(save_figs){
  ggsave(filename = here("output", "fig4_sero_tf_mda_pearson_wt.pdf"),
         fig4,
         device = pdf,
         width = 180, height = 150, units = "mm")
}
```

### Figure 5

```{r fig5, fig.width = 8.5, fig.height = 5.4}
fig5 <- plot_grid(plot_grid(sp_all_fig, sp_1to9_fig, sp_1to5_fig, nrow = 1),
                  plot_grid(tf_all_fig, tf_1to9_fig, tf_1to5_fig, nrow = 1),
                  fig4_step3,
                  ncol = 1, rel_heights = c(1, 0.9, 0.3), labels = c('a', 'b', ""))

fig5

if(save_figs){
  ggsave(filename = here("output", "fig5_sero_tf_age_pearson_wt.pdf"),
         fig5,
         device = pdf,
         width = 180, height = 150, units = "mm")
}
```

### Figure S4

```{r figS4, fig.width = 8.5, fig.height = 5.4}
figS4 <- plot_grid(plot_grid(scr_all_fig, scr_noMDA_fig, scr_MDA_fig, nrow = 1),
                  plot_grid(scr_all_fig, scr_1to9_fig, scr_1to5_fig, nrow = 1),
                  fig4_step3,
                  ncol = 1, rel_heights = c(1, 1, 0.4), labels = c('a', 'b', ""))

figS4

if(save_figs){
  ggsave(filename = here("output", "figS4_scr_pcr_pearson_wt.pdf"),
         figS4,
         device = pdf,
         width = 180, height = 150, units = "mm")
 }
```


# Session info
```{r session info}
sessionInfo()
```
